requirements.txt
python version-3.8
*libraries *
tensorflow
pillow
tkinter
numpy
pandas
spellchecker
**
system 
os
Visual studio
